Fix problems with packaging
Unity's package manager expects the meta files to exist, so they need to
be tracked by git now. Additionally, Unity will actually show the
manifest in the editor and using it will automatically reformat the
package file to use spaces, so `package.json` has been updated
accordingly.

The package type has also been set to "library" despite the online
documentation claiming that this field has been reserved for future use.
You can set this field in the inspector for the manifest and it's used
to determine if the package is visible in the editor. "library" has been
chosen as it seems like the closest thing to "framework".